action tool multiproduct adapted and jenkinsjob info produced
