+++ packages/action/package.json
@@ -44,6 +44,7 @@
     "@kie/build-chain-configuration-reader": "^2.2.6",
     "@octokit/rest": "^18.12.0",
     "argparse": "^2.0.1",
+    "axios": "^0.26.1",
     "commander": "^8.3.0",
     "dotenv": "^8.2.0",
     "fs-extra": "^9.0.0",
